🚚 Merge serialize_yaml to AssetRepository module
